+++ tests/pyccel/test_pyccel.py
@@ -863,9 +863,8 @@ def test_basic_header():
 
 #------------------------------------------------------------------------------
 @pytest.mark.parametrize( "test_file", ["scripts/classes/classes.py",
-                                        "scripts/classes/classes_1.py",
-                                        "scripts/classes/classes_5.py",
                                         "scripts/classes/generic_methods.py",
+                                        "scripts/classes/classes_1.py",
                                         ] )
 @pytest.mark.parametrize( 'language', (
         pytest.param("python", marks = pytest.mark.python),
@@ -882,6 +881,8 @@ def test_classes_f_only( test_file , language):
 #------------------------------------------------------------------------------
 @pytest.mark.xdist_incompatible
 @pytest.mark.parametrize( "test_file", ["scripts/classes/classes_2_C.py",
+                                        "scripts/classes/classes_5.py",
+                                        "scripts/classes/classes_3.py",
                                         ] )
 @pytest.mark.parametrize( 'language', (
         pytest.param("python", marks = pytest.mark.python),