+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
+-   #1472 : Add C printing support for a class containing scalar data.
+
 ### Fixed
 
 -   #1484 : Use scope for classes to avoid name clashes.